Skip SSH setup on WinRM boxes and configure WinRM last
If we configure winrm too early the box will be done while updates are still installing and the result is a non-working box so make it a zz script. This also means that it will run after update.bat with the post install updates script, which is now included along with update.bat.

Fix some bad logging in the winrm setup script that resulted in funky output

Remove the fixnetwork.ps1 script, which set all interfaces to the work profile for winrm. This isn't necessary as we set the interfaces to work in all our desktop autounattend.xml configs

Also we were setting up ssh / ssh keys on WinRM only boxes. Remove all those scripts on the WinRM boxes. This should probably be added to a single zz_setup_transport.bat script which would handle both based on a variable, but that's for another day

Signed-off-by: Tim Smith <tsmith@chef.io>